refactored ep optimiser into separate modules
