+++ autofit/graphical/__init__.py
@@ -4,7 +4,8 @@ from .declarative.collection import FactorGraphModel
 from .declarative.factor.analysis import AnalysisFactor
 from .declarative.factor.hierarchical import _HierarchicalFactor, HierarchicalFactor
 from .expectation_propagation.ep_mean_field import EPMeanField
-from .expectation_propagation.optimiser import EPOptimiser, StochasticEPOptimiser
+from .expectation_propagation.optimiser import EPOptimiser
+from .expectation_propagation import StochasticEPOptimiser
 from .factor_graphs import FactorGraph
 from .factor_graphs.factor import Factor
 from .laplace import LaplaceOptimiser, OptimisationState
