+++ autofit/graphical/expectation_propagation/__init__.py
@@ -1,4 +1,5 @@
 from .ep_mean_field import EPMeanField
 from .history import FactorHistory, EPHistory
 from .optimiser import AbstractFactorOptimiser
-from .optimiser import EPOptimiser, StochasticEPOptimiser
+from .optimiser import EPOptimiser
+from .stochastic import StochasticEPOptimiser
